ai-code-review/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
argparse==1.4.0
httpx
